2021/09-gitea-gitlab-migration/requirements.txt
giteapy
python-gitlab
requests
bs4
